+++ logs/logger.py
@@ -35,25 +35,25 @@ else:
 #filehandler=logging.FileHandler(logfile)
 filehandler.setFormatter(formatter)
 logger.addHandler(filehandler)
-logger.info(f"[PID {os.getpid()}][LOGGER START RECORD.][TTGUI v{conf.getVersion()}][LOGGIN LEVEL : {logger.level}]")
+logger.info(f"[PID {os.getpid()}]\t[LOGGER START RECORD.][TTGUI {conf.getVersion()}][LOGGIN LEVEL {logger.level}]")
 
 def debug(msg):
-    msg = f"[PID {os.getpid()}]" +msg 
+    msg = f"[PID {os.getpid()}]\t" +msg 
     logger.debug(msg)
 def info(msg):
-    msg = f"[PID {os.getpid()}]" +msg 
+    msg = f"[PID {os.getpid()}]\t" +msg 
     logger.info(msg)
 def warning(msg):
-    msg = f"[PID {os.getpid()}]" +msg 
+    msg = f"[PID {os.getpid()}]\t" +msg 
     logger.warning(msg)
 def error(msg):
-    msg = f"[PID {os.getpid()}]" +msg 
+    msg = f"[PID {os.getpid()}]\t" +msg 
     logger.error(msg)
 def critical(msg):
-    msg = f"[PID {os.getpid()}]" +msg 
+    msg = f"[PID {os.getpid()}]\t" +msg 
     logger.critical(msg)
 def log(level, msg):
-    msg = f"[PID {os.getpid()}]" +msg 
+    msg = f"[PID {os.getpid()}]\t" +msg 
     logger.log(level, msg)
 def setLevel(level):
     logger.setLevel(level)